+++ frontend_desktop/main.py
@@ -15,6 +15,7 @@ from PySide6.QtWidgets import (
     QHBoxLayout,
     QMainWindow,
     QProgressBar,
+    QPushButton,
     QStatusBar,
     QWidget,
 )
@@ -55,10 +56,23 @@ class MainWindow(QMainWindow):
         # setup status bar
         self.status_bar = QStatusBar(self)
         self.setStatusBar(self.status_bar)
+
+        # status progress bar
         self.status_progress_bar = QProgressBar(self)
         self.status_progress_bar.setFixedWidth(170)
         self.status_progress_bar.hide()
         self.status_bar.addPermanentWidget(self.status_progress_bar)
+
+        # timer for resetting UI after another click
+        self._reset_timer = QTimer(self, interval=3000)
+        self._reset_timer.timeout.connect(self._reset_tab_clicked)
+
+        # status reset button
+        self.reset_btn = QPushButton("Reset", self)
+        self.reset_btn.clicked.connect(self._reset_tab_clicked)
+        self.reset_btn.hide()
+        self.status_bar.addPermanentWidget(self.reset_btn)
+        GSigs().tab_loaded.connect(self._enable_reset_btn)
 
         # left nav (buttons only)
         self.nav = NavigationTabs(self)
@@ -204,6 +218,46 @@ class MainWindow(QMainWindow):
                     return True
 
         return super().eventFilter(obj, event)
+
+    @Slot()
+    def _reset_tab_clicked(self) -> None:
+        """Calls `reset_tab` if the timer is active. Otherwise, starts the timer and changes button text."""
+        if self._reset_timer.isActive():
+            self._reset()
+            self._stop_reset_timer()
+        else:
+            self.reset_btn.setText("Confirm?")
+            self._reset_timer.start(3000)
+
+    def _stop_reset_timer(self) -> None:
+        """Stops the reset timer if active."""
+        if self._reset_timer.isActive():
+            self._reset_timer.stop()
+        self._reset_reset_btn()
+
+    def _reset(self) -> None:
+        """Resets all tabs to their default state."""
+        for widget in self.tabs.values():
+            if hasattr(widget, "reset_tab"):
+                widget.reset_tab()
+            elif hasattr(widget, "multi_track"):
+                widget.multi_track.reset_to_single_tab()
+            else:
+                LOG.warning(
+                    f"Tab {widget} does not have a reset_tab method or multi_track attribute."
+                )
+
+        self._reset_reset_btn()
+
+    def _enable_reset_btn(self) -> None:
+        """Enables the reset button."""
+        self._stop_reset_timer()
+        self.reset_btn.show()
+
+    def _reset_reset_btn(self) -> None:
+        """Resets the reset button to its default state."""
+        self.reset_btn.setText("Reset")
+        self.reset_btn.hide()
 
 
 if __name__ == "__main__":